Fix errors reported by scala 3 compiler in javalib

+++ javalib/src/main/scala/java/io/RandomAccessFile.scala
@@ -187,7 +187,7 @@ class RandomAccessFile private (
       val fileLength = length()
       val toSkip =
         if (currentPosition + n > fileLength) fileLength - currentPosition
-        else n
+        else n.toLong
       seek(toSkip)
       toSkip.toInt
     }
